requirements.txt
streamlit
google-cloud-aiplatform
googletrans==4.0.0-rc1
legacy-cgi
Pillow